[added] Add thematic folders and move files
