+++ Makefile
@@ -6,11 +6,13 @@ MLX_LIB		= libmlx.a
 # MLX			= minilibx-linux
 # MLX_LIB		= libmlx_Linux.a
 SRCS		= minirt.c \
-			file_read.c \
-			camera.c \
-			vector.c \
-			util.c \
-			draw.c
+			_parse/file_read.c \
+			_raytracing/camera.c \
+			vector/vector.c \
+			_utils/util.c \
+			_draw/draw.c \
+			parse/extension.c \
+
 OBJS 		= $(SRCS:.c=.o)
 OBJS = $(addprefix obj/,$(SRCS:.c=.o))
 OBJS_BONUS = $(addprefix obj/,$(SRCS_BONUS:.c=.o))
