+++ _draw/cylinder.c
@@ -1,20 +1,13 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   camera.c                                           :+:      :+:    :+:   */
+/*   cylinder.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nahyulee <nahyulee@student.42seoul.kr>     +#+  +:+       +#+        */
+/*   By: soohkang <soohkang@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/22 01:25:16 by nahyulee          #+#    #+#             */
-/*   Updated: 2023/11/25 22:09:28 by nahyulee         ###   ########.fr       */
+/*   Created: 2023/12/04 17:44:57 by soohkang          #+#    #+#             */
+/*   Updated: 2023/12/04 17:44:59 by soohkang         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minirt.h"
-
-void	move_camera(t_c camera, int x, int y, int z)
-{
-	camera.cam.x += x;
-	camera.cam.y += y;
-	camera.cam.z += z;
-}
+#include "../minirt.h"
