+++ _parse/file_read.c
@@ -6,21 +6,19 @@
 /*   By: soohkang <soohkang@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/17 10:16:44 by nahyulee          #+#    #+#             */
-/*   Updated: 2023/12/04 15:36:41 by soohkang         ###   ########.fr       */
+/*   Updated: 2023/12/04 17:45:11 by soohkang         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minirt.h"
-#include "mlx/mlx.h"
+#include "../minirt.h"
+#include "../mlx/mlx.h"
 
-void	check_data_condition(t_rt *rt, char *tmp);
-void	put_r(t_rt *rt, char *tmp);
-void	put_a(t_rt *rt, char *tmp);
-
+// void	read_data(char **av)
 void	read_data(t_rt *rt, char **av)
 {
-	char	*tmp;
-	int		fd;
+	// (void) rt;
+	// char	*tmp;
+	// int		fd;
 	int		i;
 
 	i = 1;
@@ -28,18 +26,20 @@ void	read_data(t_rt *rt, char **av)
 		i++;
 	if (ft_strnstr(av[i], ".rt", ft_strlen(av[i])) == NULL)
 		ft_exit(1, "Error\nFile is not .rt\n");
+	
 
 	rt->height = 300; // 창 가로
 	rt->width = 300; // 창 세로
+	sphere(rt);
 	
-	fd = open(av[1], O_RDONLY);
-	tmp = get_next_line(fd);
-	while (*tmp)
-	{
-		check_data_condition(rt, tmp);
-		free(tmp);
-		tmp = get_next_line(fd);
-	}
+	// fd = open(av[i], O_RDONLY);
+	// tmp = get_next_line(fd);
+	// while (*tmp)
+	// {
+	// 	check_data_condition(rt, tmp);
+	// 	free(tmp);
+	// 	tmp = get_next_line(fd);
+	// }
 }
 
 void	check_data_condition(t_rt *rt, char *tmp)
@@ -111,4 +111,17 @@ unsigned int	rgb_hex(int red, int green, int blue)
 
 	color = (red << 16) | (green << 8) | blue;
 	return (color);
+}
+
+void	sphere(t_rt *rt)
+{
+	// 구체를 고정값으로 처리한 상태. 즉, 파싱해서 이 고정된 값에 다가 값을 넣어야 함
+	// sphere: 구체
+	rt->data.sphere.center = (t_vtr3){0, 0, 20};
+	rt->data.sphere.radius = 20;
+	rt->data.sphere.color = rgb_hex(255, 0, 0);
+	
+	rt->data.camera.cam = (t_vtr3){-50, 0, 20};
+	rt->data.camera.dir = (t_vtr3){0, 0, 0};
+	rt->data.camera.fov = 70;
 }
