+++ _raytracing/camera.c
@@ -1,0 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   camera.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: soohkang <soohkang@student.42seoul.kr>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/22 01:25:16 by nahyulee          #+#    #+#             */
+/*   Updated: 2023/12/04 17:46:17 by soohkang         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../minirt.h"
+
+void	move_camera(t_c camera, int x, int y, int z)
+{
+	camera.cam.x += x;
+	camera.cam.y += y;
+	camera.cam.z += z;
+}
